third_party/blink/web_tests/fast/animation/scroll-animations/scroll-timeline-snapshotting.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ScrollTimeline snapshotting</title>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#event-loop">
<link rel="help" href="https://github.com/whatwg/html/pull/11613">
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<script src='resources/testcommon.js'></script>
<style>
  body {
    height: 800px;
    width: 800px;
  }
</style>
<div id="log"></div>

<script>
'use strict';

promise_test(async t => {
  internals.settings.setScrollAnimatorEnabled(true);

  const scroller = document.scrollingElement;
  const maxScroll = scroller.scrollHeight - scroller.clientHeight;
  const timeline = new ScrollTimeline({scrollSource: scroller});

  if (!window.eventSender) {
    document.write('This test requires eventSender');
    return;
  }

  eventSender.keyDown("ArrowDown");
  // Wait for the scroll to change.
  await waitForEvent(document, "scroll");
  // The next frame the scroll changes, the scroll-timelines should not be
  // updated before the scroll events are dispatched. Store the scroll position
  // from the previous frame to compare the timeline to.
  let scroll_percentage = (scroller.scrollTop / maxScroll) * 100;

  await waitForEvent(document, "scroll");

  assert_percents_approx_equal(
      timeline.currentTime,
      scroll_percentage,
      maxScroll,
      'Scroll timeline current time corresponds to the scroll position.');
}, 'Scroll timeline current time is updated after handling user-driven scroll.');
</script>
